Remove srpt build for debian 6 builditf should not fail due to srpt compilation failures on RHEL 5, SLES SP 1 etc Remove ubuntu11 from buildinit. We never really supported ubuntu

+++ scripts/builditf.linux.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+checkerrorwarnonly() {
+	if [ "$?" != "0" ]; then
+		echo "ERROR: Building one or more kernel modules failed!. Skipping and continuing."
+	fi
+}
+
 checkerror() {
 	if [ "$?" != "0" ]; then
 		echo "ERROR: Building kernel modules failed!"
@@ -47,9 +53,11 @@ cp -f qla2xxx.ko /quadstor/lib/modules/$kvers/
 
 cd /quadstor/src/target-mode/fc/srpt
 make clean && make 
-checkerror
+checkerrorwarnonly
 
-cp -f ib_srpt.ko /quadstor/lib/modules/$kvers/
+if [ -f ib_srpt.ko ]; then
+	cp -f ib_srpt.ko /quadstor/lib/modules/$kvers/
+fi
 
 #Install the newly build qla2xxx driver
 /quadstor/bin/qlainst
